Fixed model to api mismatches

+++ backend/apis/project_apis/Manage_milestone_apis.py
@@ -84,7 +84,7 @@ class MilestonesResource(Resource):
 
 @api_bp_milestones.route("/api/milestones/<int:milestone_id>")
 class MilestoneResource(Resource):
-    #@jwt_required()
+    @jwt_required()
     @api_bp_milestones.response(200, MilestoneUpdateResponse)
     def put(self, milestone_id):
         """Update an existing milestone"""
@@ -93,18 +93,18 @@ class MilestoneResource(Resource):
             current_user = User.query.get_or_404(current_user_id)
 
             allowed_roles = ["Admin", "TA", "Instructor", "Developer"]
-            #if current_user.user_type not in allowed_roles:
-            #    return {"message": "You do not have permission to update milestone"}, 403
+            if current_user.user_type not in allowed_roles:
+                return {"message": "You do not have permission to update milestone"}, 403
 
             schema = MilestoneUpdateSchema()
             data = schema.load(request.get_json())  # Validate the incoming data
 
             milestone = Milestone.query.get_or_404(milestone_id)
 
-            if "name" in data:
-                milestone.name = data["name"]
+            if "milestone_name" in data:
+                milestone.milestone_name = data["milestone_name"]
             if "description" in data:
-                milestone.description = data["description"]
+                milestone.milestone_description = data["milestone_description"]
             if "start_date" in data:
                 milestone.start_date = datetime.strptime(data["start_date"], "%Y-%m-%d")
             if "end_date" in data:
@@ -131,8 +131,8 @@ class MilestoneResource(Resource):
             current_user = User.query.get_or_404(current_user_id)
 
             allowed_roles = ["Admin", "TA", "Instructor", "Developer"]
-            #if current_user.user_type not in allowed_roles:
-            #    return {"message": "You do not have permission to delete milestones"}, 403
+            if current_user.user_type not in allowed_roles:
+                return {"message": "You do not have permission to delete milestones"}, 403
 
             milestone = Milestone.query.get_or_404(milestone_id)
             db.session.delete(milestone)
